# contracts/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/modules/{idOrSlug}/commands/identify": {
      "post": {
        "operationId": "ModuleDiagnosticsController_identify",
        "parameters": [
          {
            "name": "idOrSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentifyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ordena al módulo identificarse (parpadeo) para localizarlo",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/modules/{idOrSlug}/targets/{targetIndex}/test-led": {
      "post": {
        "operationId": "ModuleDiagnosticsController_testLed",
        "parameters": [
          {
            "name": "idOrSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetIndex",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LedTestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Prueba de LED de una diana",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/modules/{idOrSlug}/targets/{targetIndex}/test-sensor": {
      "post": {
        "operationId": "ModuleDiagnosticsController_testSensor",
        "parameters": [
          {
            "name": "idOrSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetIndex",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestIdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pide el autodiagnóstico del módulo (no hay prueba por diana en v1)",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/modules/{idOrSlug}/targets/{targetIndex}/calibrate": {
      "post": {
        "operationId": "ModuleDiagnosticsController_calibrate",
        "parameters": [
          {
            "name": "idOrSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "targetIndex",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestIdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Arranca la calibración de una diana",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/modules/{idOrSlug}/commands/abort-calibration": {
      "post": {
        "operationId": "ModuleDiagnosticsController_abortCalibration",
        "parameters": [
          {
            "name": "idOrSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestIdDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Aborta la calibración en curso (categoría \"seguridad\": se acepta siempre)",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/modules/{idOrSlug}/diagnostics": {
      "get": {
        "operationId": "ModuleDiagnosticsController_results",
        "parameters": [
          {
            "name": "idOrSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resultados que el módulo ha devuelto de verdad",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/modules": {
      "get": {
        "operationId": "ModulesController_list",
        "parameters": [
          {
            "name": "skip",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista module",
        "tags": [
          "modules"
        ]
      },
      "post": {
        "operationId": "ModulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crea module",
        "tags": [
          "modules"
        ]
      }
    },
...
    "/api/maintenance/incidents": {
      "get": {
        "operationId": "MaintenanceController_incidents",
        "parameters": [
          {
            "name": "severity",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Incidencias registradas (rechazos de ingesta, baja tensión, OTA…)",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/api/maintenance/incidents/{id}/resolve": {
      "patch": {
        "operationId": "MaintenanceController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "maintenance"
        ]
      }
    },
    "/api/maintenance/modules/{slug}/maintenance": {
      "post": {
        "operationId": "MaintenanceController_setMaintenance",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activa o desactiva el modo mantenimiento de un módulo",
        "tags": [
          "maintenance"
        ]
      }
    },
    "/api/maintenance/modules/{slug}/self-test": {
      "post": {
        "operationId": "MaintenanceController_selfTest",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "maintenance"
        ]
      }
    },
    "/api/maintenance/modules/{slug}/identify": {
      "post": {
        "operationId": "MaintenanceController_identify",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "maintenance"
        ]
      }
    },
...
  },
  "info": {
    "title": "Diana · API del servidor de dianas modulares",
    "description": "API REST del backend Diana (WP-02).\n\nNotas normativas:\n- Los tiempos de impacto se exponen en MICROSEGUNDOS y por separado: `device_event_us` (T1, del ESP32), `coordinator_elapsed_us` (T2, del módulo principal), `received_at` (T3) y `persisted_at` (T4). El backend nunca reescribe T1 ni T2 (ADR-0002).\n- `accuracy_status` puede valer `not_computable`: en ese caso `shots_fired`, `accuracy_total` y `accuracy_valid` son `null` y NO deben mostrarse como 0 % (ADR-0006).\n- Los duplicados de impacto son normales con QoS 1 y no son errores (ADR-0003).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "IdentifyDto": {
        "type": "object",
        "properties": {}
      },
      "LedTestDto": {
        "type": "object",
        "properties": {}
      },
      "RequestIdDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}